requirements.txt
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dpcpp-cpp-rt==2025.3.2
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.2.0
ImageIO==2.37.3
iniconfig==2.3.0
intel-cmplr-lib-rt==2025.3.2
intel-cmplr-lib-ur==2025.3.2
intel-cmplr-lic-rt==2025.3.2
intel-opencl-rt==2025.3.2
intel-openmp==2025.3.2
intel-pti==0.16.0
intel-sycl-rt==2025.3.2
Jinja2==3.1.6
kiwisolver==1.5.0
lazy-loader==0.5
MarkupSafe==3.0.3
matplotlib==3.10.9
mkl==2025.3.1
monai==1.5.2
mpmath==1.3.0
networkx==3.6.1
nibabel==5.4.2
numpy==2.4.4
onemkl-license==2025.3.1
onemkl-sycl-blas==2025.3.1
onemkl-sycl-dft==2025.3.1
onemkl-sycl-lapack==2025.3.1
onemkl-sycl-rng==2025.3.1
onemkl-sycl-sparse==2025.3.1
packaging==26.1
pillow==12.2.0
pluggy==1.6.0
pyelftools==0.32
Pygments==2.20.0
pyparsing==3.3.2
pytest==9.0.3
python-dateutil==2.9.0.post0
PyYAML==6.0.3
scikit-image==0.26.0
scipy==1.17.1
setuptools==70.2.0
six==1.17.0
sympy==1.14.0
tbb==2022.3.1
tcmlib==1.4.1
tifffile==2026.4.11
torch==2.11.0+xpu
tqdm==4.67.3
triton-xpu==3.7.0
typing_extensions==4.15.0
umf==1.0.3